Refactored Javascripts
Consolidated all the javascripts into 2 files. There is model.js, which is the high level JS. It manages the Statistics themselves and then calls each individual function to edit the DOM accordingly. Then there is visuals.js, which is full of independent functions that edit the DOM. These should not call each other, but instead be handled by the controller. This code is subject to refactoring into a better abstractions. But for now, tests pass.
