+++ smap/package.json
@@ -15,7 +15,8 @@
     "jade": "~1.11.0",
     "jquery": "^3.4.1",
     "morgan": "~1.9.1",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "rewire": "^4.0.1"
   },
   "devDependencies": {
     "jest": "^25.1.0"
