+++ smap/tests/client/statistics.test.js
@@ -3,7 +3,8 @@
 
 "use strict";
 
-var $ = require('jquery');
+const $ = require('jquery');
+const rewire = require('rewire');
 
 /*
   I tried so many ways to get external scripts to load correctly.
@@ -15,7 +16,7 @@ var $ = require('jquery');
   External citation: https://info340.github.io/jest.html
 */
 window.$ = $;
-window.mix_color = jest.fn(weight => "rgba(0,0,0,1)");
+// window.mixColor = jest.fn(weight => "rgba(0,0,0,1)");
 
 //Sample test for now
 test('Create a slider', (done) => {
@@ -32,8 +33,18 @@ test('Create a slider', (done) => {
     <div id=statistics-selector><div id="inactive_slider_template" class="statistic-option">Job Openings</div></div>
   `;
 
+  let inactive_slider = $("#inactive_slider_template");
+  let active_slider = $("#active_slider_template");
   //Has side effects, so must be included in the test
   let script = require("../../public/javascripts/statistics");
+
+  // Note: Check if function is called
+  window.colorState = jest.fn( () => {} );
+  window.makeActiveSlider = jest.fn( () => { return active_slider; } );
+  window.makeInactiveSlider = jest.fn( () => { return inactive_slider; } );
+  // script.__set__("displayWeights", () => {});
+  // var spy = jest.spyOn(script, "displayWeights").mockImplementation( () => {} );
+
 
   //The ready function. Since these callbacks are called in order,
   //  the one in statistics.js is called first to set up the event listener.
@@ -43,10 +54,12 @@ test('Create a slider', (done) => {
     try {
       let stat = new script.Stat({title:"test stat", id:0}, script.DEFAULT_WEIGHT);
       expect($(".statistic-option").length).toEqual(1);
-      expect($(".statistic-slider-container").length).toEqual(0);
-      $(".statistic-option").click();
+      expect(window.makeActiveSlider.mock.calls.length).toBe(0);
+      // expect($(".statistic-slider-container").length).toEqual(0);
+      inactive_slider.click();
       expect($(".statistic-option").length).toEqual(0);
-      expect($(".statistic-slider-container").length).toEqual(1);
+      expect(window.makeActiveSlider.mock.calls.length).toBe(1);
+      // expect($(".statistic-slider-container").length).toEqual(1);
       done();
     } catch (error) {
       done(error);
